Moved validation fixes to new files. Removed fednlib.cpp
